requirements.txt
elevenlabs
pyaudio
openai
python-dotenv
pillow
langchain_community
#wave

#offline mode dependencies
openai-whisper
pyttsx3
apscheduler
requests 
pynput
keyboard
word2number
werkzeug
flask
pyinstaller
